refactor(policies): suffix OpenAI provider policies with transformer
