+++ docs/README.md
@@ -59,11 +59,11 @@ All available policies, sorted alphabetically.
 | [Model Weighted Round Robin](./model-weighted-round-robin/v1.0/docs/model-weighted-round-robin.md) | AI | Implements weighted round-robin load balancing for AI models. |
 | [NeMo Guard Content Safety](./nvidia-nemoguard-content-safety/v0.9/docs/nvidia-nemoguard-content-safety.md) | Guardrails, AI | Validates request and/or response content using NVIDIA NeMo Guard (llama-3.1-nemoguard-8b-content-safety). |
 | [Opaque Token Auth](./opaque-token-auth/v1.0/docs/opaque-token-authentication.md) | Security, AI | Validates opaque OAuth 2.0 access tokens via RFC 7662 token introspection. |
-| [OpenAI to Anthropic](./openai-to-anthropic/v0.9/docs/openai-to-anthropic.md) | AI | Translates OpenAI Chat Completions requests into Anthropic Messages API format and rewrites non-streaming Anthropic responses back into the OpenAI response shape. |
-| [OpenAI to Azure OpenAI](./openai-to-azure-openai/v0.9/docs/openai-to-azure-openai.md) | AI | Routes OpenAI Chat Completions requests to an Azure OpenAI upstream by rewriting the request path with the deployment id and api-version query parameter. |
-| [OpenAI to Gemini](./openai-to-gemini/v0.9/docs/openai-to-gemini.md) | AI | Translates OpenAI Chat Completions requests into the Google Gemini generateContent API format and rewrites non-streaming Gemini responses back into the OpenAI response shape. |
-| [OpenAI to Mistral](./openai-to-mistral/v0.9/docs/openai-to-mistral.md) | AI | Adapts OpenAI Chat Completions requests for Mistral-compatible upstreams by pinning the model, stripping OpenAI fields Mistral rejects, and rewriting the request path. |
+| [OpenAI to Anthropic Transformer](./openai-to-anthropic-transformer/v0.9/docs/openai-to-anthropic-transformer.md) | AI | Translates OpenAI Chat Completions requests into Anthropic Messages API format and rewrites non-streaming Anthropic responses back into the OpenAI response shape. |
+| [OpenAI to Azure OpenAI Transformer](./openai-to-azure-openai-transformer/v0.9/docs/openai-to-azure-openai-transformer.md) | AI | Routes OpenAI Chat Completions requests to an Azure OpenAI upstream by rewriting the request path with the deployment id and api-version query parameter. |
 | [OpenAI to Bedrock Transformer](./openai-to-bedrock-transformer/v0.9/docs/openai-to-bedrock-transformer.md) | AI | Translates OpenAI Chat Completions requests into AWS Bedrock Converse format and rewrites non-streaming and streaming Bedrock responses back into OpenAI-compatible JSON or Server-Sent Events. |
+| [OpenAI to Gemini Transformer](./openai-to-gemini-transformer/v0.9/docs/openai-to-gemini-transformer.md) | AI | Translates OpenAI Chat Completions requests into the Google Gemini generateContent API format and rewrites non-streaming Gemini responses back into the OpenAI response shape. |
+| [OpenAI to Mistral Transformer](./openai-to-mistral-transformer/v0.9/docs/openai-to-mistral-transformer.md) | AI | Adapts OpenAI Chat Completions requests for Mistral-compatible upstreams by pinning the model, stripping OpenAI fields Mistral rejects, and rewriting the request path. |
 | [PII Masking Regex](./pii-masking-regex/v1.0/docs/pii-masking-regex.md) | Guardrails, AI | Masks or redacts Personally Identifiable Information (PII) from request/response bodies using regex patterns. |
 | [Prompt Compressor](./prompt-compressor/v0.9/docs/prompt-compressor.md) | AI | Compresses selected prompt text in JSON request bodies before upstream LLM calls. |
 | [Prompt Decorator](./prompt-decorator/v1.0/docs/prompt-decorator.md) | AI | Dynamically modifies the prompt by applying custom decorations using a configured strategy. |
