+++ docs/llm-header-router/v0.9/docs/llm-header-router.md
@@ -5,7 +5,7 @@ title: "Overview"
 
 ## Overview
 
-The LLM Header Router policy is the front door of a multi-provider LLM proxy. It reads a configurable request header (default `x-provider`), matches it against a list of value → provider mappings, and publishes the chosen provider id into `SharedContext.Metadata["selected_provider"]`. Downstream translator policies (for example `openai-to-anthropic`, `openai-to-azure-openai`, `openai-to-gemini`, or `openai-to-mistral`) read that key and run only when the selection matches their own `providerId`. This lets a single OpenAI-shaped `/chat/completions` endpoint fan out to many providers, chosen per request by one header, with no change to the client's request body.
+The LLM Header Router policy is the front door of a multi-provider LLM proxy. It reads a configurable request header (default `x-provider`), matches it against a list of value → provider mappings, and publishes the chosen provider id into `SharedContext.Metadata["selected_provider"]`. Downstream translator policies (for example `openai-to-anthropic-transformer`, `openai-to-azure-openai-transformer`, `openai-to-gemini-transformer`, or `openai-to-mistral-transformer`) read that key and run only when the selection matches their own `providerId`. This lets a single OpenAI-shaped `/chat/completions` endpoint fan out to many providers, chosen per request by one header, with no change to the client's request body.
 
 The selection is published in both the request-header phase and the request-body phase. The header phase runs first so header-phase consumers — such as the proxy → provider upstream auth injection — observe the selection before they evaluate their `selected_provider` gate. The body phase repeats the publish idempotently: if `selected_provider` is already set, it is left untouched.
