--- docs/openai-to-anthropic/v0.9/metadata.json
@@ -1,10 +1,0 @@
-{
-  "name": "openai-to-anthropic",
-  "displayName": "OpenAI to Anthropic",
-  "version": "0.9",
-  "provider": "WSO2",
-  "categories": [
-    "AI"
-  ],
-  "description": "Translates OpenAI Chat Completions requests into Anthropic Messages API format and rewrites non-streaming Anthropic responses back into the OpenAI response shape. Supports system prompts, multi-modal image input, tool/function calling, and both routed (multi-provider) and unconditional (single-provider) execution modes."
-}
