+++ docs/openai-to-azure-openai-transformer/v0.9/metadata.json
@@ -1,0 +1,10 @@
+{
+  "name": "openai-to-azure-openai-transformer",
+  "displayName": "OpenAI to Azure OpenAI Transformer",
+  "version": "0.9",
+  "provider": "WSO2",
+  "categories": [
+    "AI"
+  ],
+  "description": "Routes OpenAI Chat Completions requests to an Azure OpenAI upstream by rewriting the request path with the deployment id and api-version query parameter. The request body passes through unchanged, since Azure OpenAI is wire-compatible with the OpenAI Chat Completions API. Supports both routed (multi-provider) and unconditional (single-provider) execution modes."
+}
