--- docs/openai-to-gemini/v0.9/metadata.json
@@ -1,10 +1,0 @@
-{
-  "name": "openai-to-gemini",
-  "displayName": "OpenAI to Gemini",
-  "version": "0.9",
-  "provider": "WSO2",
-  "categories": [
-    "AI"
-  ],
-  "description": "Translates OpenAI Chat Completions requests into the Google Gemini generateContent API format and rewrites non-streaming Gemini responses back into the OpenAI response shape. Maps messages to contents, system prompts to systemInstruction, tools to functionDeclarations, and multi-modal image input. Supports both routed (multi-provider) and unconditional (single-provider) execution modes."
-}
