--- docs/openai-to-mistral/v0.9/metadata.json
@@ -1,10 +1,0 @@
-{
-  "name": "openai-to-mistral",
-  "displayName": "OpenAI to Mistral",
-  "version": "0.9",
-  "provider": "WSO2",
-  "categories": [
-    "AI"
-  ],
-  "description": "Adapts OpenAI Chat Completions requests for Mistral-compatible upstreams by pinning the model, stripping OpenAI fields Mistral rejects, and rewriting the request path. Mistral already emits OpenAI-shaped responses, so responses are normalised and passed back in OpenAI shape. Supports both routed (multi-provider) and unconditional (single-provider) execution modes."
-}
