+++ policies/openai-to-anthropic-transformer/response.go
