--- policies/openai-to-anthropic/go.mod
@@ -1,5 +1,0 @@
-module github.com/wso2/gateway-controllers/policies/openai-to-anthropic
-
-go 1.26.2
-
-require github.com/wso2/api-platform/sdk/core v0.2.14
