+++ policies/openai-to-gemini-transformer/go.mod
@@ -1,0 +1,5 @@
+module github.com/wso2/gateway-controllers/policies/openai-to-gemini-transformer
+
+go 1.26.2
+
+require github.com/wso2/api-platform/sdk/core v0.2.14
