--- policies/openai-to-gemini/policy-definition.yaml
@@ -1,46 +1,0 @@
-name: openai-to-gemini
-version: v0.9.0
-description: |
-  Translates OpenAI Chat Completions requests to the Gemini generateContent API
-  format and rewrites non-streaming Gemini responses back to the OpenAI shape.
-  Supports routed and unconditional execution modes.
-
-parameters:
-  type: object
-  additionalProperties: false
-  properties:
-    model:
-      type: string
-      x-wso2-policy-advanced-param: false
-      description: |
-        Gemini model name used in the translated request (for example
-        gemini-2.5-pro). Overrides the OpenAI "model" field and is used
-        in the rewritten path.
-      minLength: 1
-    providerId:
-      type: string
-      x-wso2-policy-advanced-param: false
-      description: |
-        Provider this translator targets. Used as the upstream cluster name
-        (matching an id or as alias in the LlmProxy additionalProviders list) and, in
-        multi-provider mode, as the key matched (case-insensitive) against
-        SharedContext.Metadata["selected_provider"]. When no provider has been
-        selected upstream (single-provider proxy), the translator always runs;
-        when a provider has been selected, it runs only if the selection
-        matches this provider id. When omitted, routing is left to the route's
-        default upstream.
-      minLength: 1
-    apiVersion:
-      type: string
-      x-wso2-policy-advanced-param: true
-      description: |
-        Gemini API version segment used in the rewritten path
-        (/{apiVersion}/models/{model}:generateContent).
-      default: "v1beta"
-      minLength: 1
-  required:
-    - model
-
-systemParameters:
-  type: object
-  properties: {}
