+++ policies/openai-to-mistral-transformer/response.go
